Create Code
Over the past few lessons, we learned more about arrays in loops. More specifically, different ways to manipulate them and to use algorithms in order to search for certain criteria. This has not only helped me try to be more efficient in my code, but also has tested my ability to visualize how to find a solution.